+++ src/decorators/Mutation.ts
@@ -1,13 +1,16 @@
-import { ReturnTypeFunc, BasicOptions } from "../types/decorators";
+import { ReturnTypeFunc, AdvancedOptions } from "../types/decorators";
 import { MetadataStorage } from "../metadata/metadata-storage";
 import { getHandlerInfo } from "../helpers/handlers";
 import { getTypeDecoratorParams } from "../helpers/decorators";
 
-export function Mutation(options?: BasicOptions): MethodDecorator;
-export function Mutation(returnTypeFunc: ReturnTypeFunc, options?: BasicOptions): MethodDecorator;
+export function Mutation(options?: AdvancedOptions): MethodDecorator;
 export function Mutation(
-  returnTypeFuncOrOptions?: ReturnTypeFunc | BasicOptions,
-  maybeOptions?: BasicOptions,
+  returnTypeFunc: ReturnTypeFunc,
+  options?: AdvancedOptions,
+): MethodDecorator;
+export function Mutation(
+  returnTypeFuncOrOptions?: ReturnTypeFunc | AdvancedOptions,
+  maybeOptions?: AdvancedOptions,
 ): MethodDecorator {
   const { options, returnTypeFunc } = getTypeDecoratorParams(returnTypeFuncOrOptions, maybeOptions);
   return (prototype, methodName) => {