+++ src/scalars/aliases.ts
@@ -1,5 +1,5 @@
-import { GraphQLFloat, GraphQLString, GraphQLInt, GraphQLScalarType } from "graphql";
+import { GraphQLFloat, GraphQLID, GraphQLInt, GraphQLScalarType } from "graphql";
 
 export const Int = GraphQLInt;
 export const Float = GraphQLFloat;
-export const ID = GraphQLString;
+export const ID = GraphQLID;