+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - add support for descriptions in schema (types, args, queries, etc.)
 - add support for declaring depreciation reason on object fields and queries/mutations
 
+### Fixes
+- fix scalars ID alias (GraphQLID not GraphQLString)
+
 ## v0.2.0
 ### Features
 - add support for Date type (built-in scalar)